requirements.txt
pandas==2.0.3
pycryptodome==3.18.0
PyQt6==6.5.1
